Fix circular dependency and switch to sqlite for DevelopmentConfig instances
