+++ .gitignore
@@ -1,3 +1,4 @@
 .venv
 .idea
 migrations
+*.db
