+++ config.py
@@ -3,9 +3,6 @@ import os
 
 class Config(object):
     DEBUG = False
-    PG_PASS = 'EazdJhQW4j'
-    PG_USER = 'genten'
-    PG_URL = os.environ['DATABASE_URL']
     SECRET_KEY = 'develop'
     SQLALCHEMY_DATABASE_URI = os.environ['DATABASE_URL']
 
@@ -13,4 +10,4 @@ class Config(object):
 class DevelopmentConfig(Config):
     DEVELOPMENT = True
     DEBUG = True
-    SQLALCHEMY_DATABASE_URI = 'postgresql://genten:develop@localhost/genten'
+    SQLALCHEMY_DATABASE_URI = 'sqlite:///genten.db'
