+++ run.py
@@ -1,4 +1,5 @@
-from app import app
+from app import app, db, routes
 
 if __name__ == '__main__':
+    routes.init(app, db)
     app.run()
